컴포넌트 UI 분리
