+++ package.json
@@ -14,7 +14,9 @@
     "date-fns": "^2.29.1",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
+    "react-redux": "^8.0.2",
     "react-scripts": "5.0.1",
+    "redux": "^4.2.0",
     "styled-components": "^5.3.5",
     "web-vitals": "^2.1.0"
   },
